requirements.txt
pylint
pytest
python-dotenv
boto3
mypy-boto3-s3
psycopg2
pandas
ipykernel
confluent-kafka